+++ dart/lib/src/seed.dart
@@ -6,7 +6,7 @@
 import 'frb_generated.dart';
 import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 
-// These functions are ignored because they are not marked as `pub`: `build_header`, `decrypt_seed_internal`, `derive_key`, `encrypt_seed_internal`
+// These functions are ignored because they are not marked as `pub`: `build_header`, `decrypt_seed_internal`, `derive_key`, `encrypt_seed_internal`, `validate_kdf_params`
 // These function are ignored because they are on traits that is not defined in current crate (put an empty `#[frb]` on it to unignore): `assert_fields_are_eq`, `clone`, `clone`, `eq`, `fmt`, `fmt`
 
 /// The recommended default Argon2id parameters for this build.